Changed unsingned int to int everywhere and fixed bugs in movement function orders

+++ src/Inputs.cpp
@@ -33,16 +33,16 @@ int Inputs::loadFromFile() {
     }
 
     // Parse each line of the input file into the variable it corresponds to
-    this->num_lanes           = (unsigned int) std::stoi(parseLine(input_lines[0]));
-    this->length              = (unsigned int) std::stoi(parseLine(input_lines[1]));
-    this->percent_full        =                std::stod(parseLine(input_lines[2]));
-    this->max_speed           = (unsigned int) std::stoi(parseLine(input_lines[3]));
-    this->look_forward        = (unsigned int) std::stoi(parseLine(input_lines[4]));
-    this->look_other_forward  = (unsigned int) std::stoi(parseLine(input_lines[5]));
-    this->look_other_backward = (unsigned int) std::stoi(parseLine(input_lines[6]));
-    this->prob_slow_down      =                std::stod(parseLine(input_lines[7]));
-    this->prob_change         =                std::stod(parseLine(input_lines[8]));
-    this->max_time            = (unsigned int) std::stoi(parseLine(input_lines[9]));
+    this->num_lanes           = std::stoi(parseLine(input_lines[0]));
+    this->length              = std::stoi(parseLine(input_lines[1]));
+    this->percent_full        = std::stod(parseLine(input_lines[2]));
+    this->max_speed           = std::stoi(parseLine(input_lines[3]));
+    this->look_forward        = std::stoi(parseLine(input_lines[4]));
+    this->look_other_forward  = std::stoi(parseLine(input_lines[5]));
+    this->look_other_backward = std::stoi(parseLine(input_lines[6]));
+    this->prob_slow_down      = std::stod(parseLine(input_lines[7]));
+    this->prob_change         = std::stod(parseLine(input_lines[8]));
+    this->max_time            = std::stoi(parseLine(input_lines[9]));
 
     // Close the input file
     input_file.close();
